+++ lib/brando/pages/pages.ex
@@ -3,7 +3,7 @@ defmodule Brando.Pages do
   Context for pages
   """
   use BrandoAdmin, :context
-  use Brando.Query
+  use Brando.Query.Compiler
 
   import Ecto.Query